Cloud-Engineering/Assignments/hw3/423-2024-hw3-scg1143-main/requirements.txt
streamlit==1.34.0
boto3==1.26.132
pyaml==24.4.0
scikit-learn==1.3.2